+++ product-service/db/sql/products.ddl.sql
